// src/app/models/itemAcervo.ts
export default interface ItemAcervo {
  codItem: string;
  codLocal: string;
  numExemplar: string;
  nomeItem: string;
  codAutor: string;
  nomeAutor: string;
  codEditora: string;
  nomeEditora: string;
  tipoItem: string;
  nomeLocal: string;
  volItem: string;
  anoEditItem: string;
  localizacaoItem: string;
  secaoItem: string;
}